♻️ move Observable and BufferedObservable to @datadog/js-core/util
Both classes have no browser-specific dependencies.
queueMicrotask moves to js-core as a pure scheduling helper (no monitor
wrapping — callers are responsible for error handling at their layer).

browser-core/tools/observable.ts becomes a thin re-export.
browser-core/tools/queueMicrotask.ts is deleted: its only consumer was
observable.ts, which now lives in js-core.
